requirements/base.txt
Package                Version
---------------------- -------
asgiref                     3.5.2
backports.zoneinfo          0.2.1
Django                      4.1
django-admin-numeric-filter 0.1.6
django-phonenumber-field    7.0.0
django-ranged-response      0.2.0
django-simple-captcha       0.5.17
phonenumbers                8.12.56
Pillow                      9.2.0
pip                         20.0.2
pkg-resources               0.0.0
psycopg2                    2.9.3
python-dotenv               0.20.0
setuptools                  44.0.0
sqlparse                    0.4.2
